requirements.txt
# Only install what's not already available system-wide
pytesseract>=0.3.10
google-cloud-vision>=3.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
ollama>=0.1.0
openai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# CLI and file processing
watchdog>=3.0.0
rich>=13.0.0
reportlab>=4.0.0

# Testing and debugging
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0
pytest-asyncio>=0.21.0

# Code quality and linting
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0
mypy>=1.11.0
ruff>=0.4.0

# Type stubs for mypy
types-requests>=2.32.0
types-PyYAML>=6.0.0
types-psutil>=7.0.0

# PyYAML==6.0.1 - already installed system-wide
# click==8.1.6 - already installed system-wide  
# sqlite3 - built into Python
# tesseract - system binary at /usr/bin/tesseract
# ollama - system binary at /usr/local/bin/ollama